testing funcs

+++ include/wavelet.hpp
@@ -1,6 +1,6 @@
 #include "imports.hpp"
 #include "common.hpp"
 
-vec1d<float> generateRicker(uint16_t Nt, float dt, uint16_t fmax);
+vec1d<float> generateRicker(uint16_t Nt, float dt, float fmax);
 
 vec1d<float> generateOrmsby(uint16_t Nt, float dt, const vec1d<float>& freqs);
